Refactored examples to all be in examples package. Updated main pom.xml to exclude impl, examples and tutorial packages. Updated build and publish doc scripts to build and publish javadocs for release.
