+++ build_docs.sh
@@ -5,7 +5,11 @@
 cd docs
 make html
 cd ..
+
 # spring boot docs
 cd extensions/spring/boot/docs
 make html
 cd ../../../../
+
+# javadocs
+mvn javadoc:aggregate -P docs
