--- examples/spring-boot-default/src/main/java/example/RestrictedErrorController.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2015 Stormpath, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package example;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.web.ErrorAttributes;
-import org.springframework.boot.autoconfigure.web.ErrorController;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.context.request.ServletRequestAttributes;
-
-import javax.servlet.http.HttpServletRequest;
-import java.util.Map;
-
-/**
- * @since 1.0.RC5
- */
-@Controller
-public class RestrictedErrorController implements ErrorController {
-    private static final String ERROR_PATH = "/error";
-
-    @Autowired
-    private ErrorAttributes errorAttributes;
-
-    @Override
-    public String getErrorPath() {
-        return ERROR_PATH;
-    }
-
-    @RequestMapping(ERROR_PATH)
-    String error(HttpServletRequest request, Model model) {
-        Map<String, Object> errorMap = errorAttributes.getErrorAttributes(new ServletRequestAttributes(request), false);
-        model.addAttribute("errors", errorMap);
-        return "error";
-    }
-}
