+++ publish_docs.sh
@@ -7,14 +7,27 @@ git clone git@github.com:stormpath/stormpath.github.io.git
 cd stormpath.github.io
 git fetch origin source:source
 git checkout source
+
+echo "Copying over servlet plugin docs"
 rm -rf source/java/servlet-plugin/
 cp -r ../docs/build/html source/java/servlet-plugin
 cp -r ../docs/build/html source/java/servlet-plugin/latest
 cp -r ../docs/build/html source/java/servlet-plugin/$RELEASE_VERSION
+
+echo "Copying over spring boot docs"
 rm -rf source/java/spring-boot-web
 cp -r ../extensions/spring/boot/docs/build/html source/java/spring-boot-web
 cp -r ../extensions/spring/boot/docs/build/html source/java/spring-boot-web/latest
 cp -r ../extensions/spring/boot/docs/build/html source/java/spring-boot-web/$RELEASE_VERSION
+
+echo "Copying over javadocs"
+rm -rf source/java/apidocs
+cp -r ../target/site/apidocs source/java/apidocs
+cp -r ../target/site/apidocs source/java/apidocs/latest
+cp -r ../target/site/apidocs source/java/apidocs/$RELEASE_VERSION
+
+ls -la source/java/apidocs/com/stormpath/spring
+
 git add --all
 git commit -m "stormpath-sdk-java release $RELEASE_VERSION"
 ls -la source/java/servlet-plugin
